Update code for new features

+++ ex1_23030327/ex1_1.java
@@ -2,8 +2,8 @@ package ex1_23030327;
 import java.util.Scanner;
 //编写程序求 n!
 public class ex1_1 {
-    public static int jc(int n) {
-        int m=0;
+    public static long jc(long n) {
+        long m=0;
         if(n==1){
             m=1;
         }else{
@@ -13,7 +13,7 @@ public class ex1_1 {
     }
     public static void main(String[] args) {
         System.out.print("请输入一个数字：");
-        int n=new Scanner(System.in).nextInt();
+        long n=new Scanner(System.in).nextInt();
         System.out.println("该数字的阶乘为："+jc(n));
     }
 }
